Split NinSnes Intelli sequence code
