+++ src/main/CMakeLists.txt
@@ -223,6 +223,7 @@ target_sources(vgmtranscore
     formats/NinSnes/NinSnesScanner.cpp
     formats/NinSnes/NinSnesScannerPatterns.cpp
     formats/NinSnes/NinSnesSeq.cpp
+    formats/NinSnes/NinSnesSeqIntelli.cpp
     formats/Org/OrgScanner.cpp
     formats/Org/OrgSeq.cpp
     formats/PS1/PS1Seq.cpp
